Fix "prctl" calls to match signature
This removes some compilation errors noticed by the Android
NDK compiler.

+++ bbque/application_proxy.cc
@@ -1449,7 +1449,7 @@ void ApplicationProxy::Dispatcher() {
 	pchMsg_t pmsg;
 
 	// Set the module name
-	if (prctl(PR_SET_NAME, BBQUE_MODULE_NAME("ap"), 0, 0, 0) != 0) {
+	if (prctl(PR_SET_NAME, (long unsigned int)BBQUE_MODULE_NAME("ap"), 0, 0, 0) != 0) {
 		logger->Error("Set name FAILED! (Error: %s)\n", strerror(errno));
 	}
