+++ bbque/platform_proxy.cc
@@ -89,7 +89,7 @@ void PlatformProxy::Monitor() {
 	std::unique_lock<std::mutex> trdStatus_ul(trdStatus_mtx);
 
 	// Set the module name
-	if (prctl(PR_SET_NAME, BBQUE_MODULE_NAME("pp"), 0, 0, 0) != 0) {
+	if (prctl(PR_SET_NAME, (long unsigned int)BBQUE_MODULE_NAME("pp"), 0, 0, 0) != 0) {
 		logger->Error("Set name FAILED! (Error: %s)\n", strerror(errno));
 	}
